feat(backend): refactor question controller
