+++ backend/src/modules/core/index.ts
@@ -9,7 +9,16 @@ import { NotificationController } from '../notification/controllers/Notification
 import { NOTIFICATION_VALIDATORS } from '../notification/validators/NotificationValidators.js';
 import { AnswerController } from '../answer/controllers/AnswerController.js';
 import { ANSWER_VALIDATORS } from '../answer/classes/validators/AnswerValidator.js';
-import { QuestionController } from '../question/controllers/QuestionController.js';
+import {
+  QuestionController,
+  QuestionIngestionController,
+  QuestionAllocationController,
+  QuestionFeedbackController,
+  QuestionPaeValidationController,
+  QuestionReportController,
+  QuestionAiController,
+  QuestionMaintenanceController,
+} from '../question/controllers/index.js';
 import { QUESTION_VALIDATORS } from '../question/classes/validators/QuestionVaidators.js';
 import { ContextController } from '../../modules/context/controllers/ContextController.js';
 import { CONTEXT_VALIDATORS } from '../context/classes/validators/ContextValidator.js';
@@ -21,6 +30,13 @@ import { REQUEST_VALIDATORS } from '../request/classes/validators/RequestValidat
 
 // Export names that loadAppModules expects
 export const coreModuleControllers: Function[] = [
+  QuestionReportController,
+  QuestionAllocationController,
+  QuestionFeedbackController,
+  QuestionPaeValidationController,
+  QuestionAiController,
+  QuestionIngestionController,
+  QuestionMaintenanceController,
   QuestionController,
   ContextController,
   AnswerController,
@@ -52,7 +68,7 @@ export const coreModuleValidators: Function[] = [
 ];
 
 // Export all the main components for external use
-export * from '../../modules/question/controllers/QuestionController.js';
+export * from '../../modules/question/controllers/index.js';
 export * from '../../modules/answer/controllers/AnswerController.js'
 export * from '../../modules/context/controllers/ContextController.js';
 export * from '../../modules/comment/controllers/CommentController.js';
