+++ backend/src/modules/question/controllers/index.ts
@@ -1,0 +1,10 @@
+export * from './QuestionController.js';
+export * from './QuestionIngestionController.js';
+export * from './QuestionAllocationController.js';
+export * from './QuestionFeedbackController.js';
+export * from './QuestionPaeValidationController.js';
+export * from './QuestionReportController.js';
+export * from './QuestionAiController.js';
+export * from './QuestionMaintenanceController.js';
+export * from './helpers/questionAuditHelper.js';
+export * from './helpers/fileUploadParser.js';
