+++ backend/src/modules/question/index.ts
@@ -1,0 +1,3 @@
+export * from './controllers/index.js';
+export * from './services/QuestionService.js';
+export * from './interfaces/index.js';
